Crate head

Enums

Flags
We truncate the input at either some number of lines or bytes

Input
Input is either a list of one or more files, or STDIN

Constants

DEFAULT_LINES_COUNT

NEW_LINE

Functions

head
Return the head of our input, truncated at a number of lines or bytes

main

read_stream
Read from a stream, truncated at a number of lines or bytes and write back to a stream